Various windows changes and an attempt at windows cmake build

+++ Graphics/ExtraBoard.cpp
@@ -8,7 +8,8 @@ ExtraBoard::ExtraBoard(float x_, float y_) {
   int mid = floor((FINALROOM+START)*1.0/2);
   for (int i=START;i<FINALROOM;i++) {
     sf::RectangleShape shape(sf::Vector2f(40,40));
-    shape.setPosition(x+40*((i-START)%mid),y+40*floor((i-START)/mid));
+		
+    shape.setPosition(x+40*((i-START)%mid),y+40*floor(double((i-START)/mid)));
     shape.setFillColor(sf::Color(0,0,0,0));
     shape.setOutlineColor(sf::Color(0,0,0));
     shape.setOutlineThickness(1);
@@ -24,7 +25,7 @@ void ExtraBoard::reset(You* you) {
       continue;
     float rad = 25/2.0;
     sf::CircleShape shape(rad);
-    shape.setPosition(x+20-rad+40*((i-START)%mid),y+20-rad+40*floor((i-START)/mid));
+    shape.setPosition(x+20-rad+40*((i-START)%mid),y+20-rad+40*floor(double(i-START)/mid));
     shape.setFillColor(sf::Color(255,0,255));
     extras.push_back(shape);
   }
